# frontend/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PromptRepo API",
    "description": "Backend API for PromptRepo application with GitHub OAuth",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/tools/": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "List all tools",
        "description": "Retrieve a list of all tools in a repository.",
        "operationId": "list_tools_api_v0_tools__get",
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Repository name",
              "default": "default",
              "title": "Repo Name"
            },
            "description": "Repository name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ToolSummary__"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to retrieve tools"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Create or update tool",
        "description": "Create a new tool or update an existing tool definition.",
        "operationId": "create_tool_api_v0_tools__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateToolRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_ToolSaveResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/validation-error",
                  "title": "Validation Error",
                  "detail": "Tool validation failed"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to create tool"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/tools/{tool_name}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get tool definition",
        "description": "Retrieve a specific tool definition by name.",
        "operationId": "get_tool_api_v0_tools__tool_name__get",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          },
          {
            "name": "repo_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Repository name",
              "default": "default",
              "title": "Repo Name"
            },
            "description": "Repository name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_ToolDefinition_"
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Not Found",
                  "detail": "Tool not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to retrieve tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tools"
        ],
        "summary": "Delete tool",
        "description": "Delete a tool definition from the repository.",
        "operationId": "delete_tool_api_v0_tools__tool_name__delete",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          },
          {
            "name": "repo_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Repository name",
              "default": "default",
              "title": "Repo Name"
            },
            "description": "Repository name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_dict_"
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Not Found",
                  "detail": "Tool not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to delete tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/tools/{tool_name}/validate": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Validate tool definition",
        "description": "Validate a tool definition for correctness.",
        "operationId": "validate_tool_api_v0_tools__tool_name__validate_post",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          },
          {
            "name": "repo_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Repository name",
              "default": "default",
              "title": "Repo Name"
            },
            "description": "Repository name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/validation-error",
                  "title": "Validation Error",
                  "detail": "Tool validation failed"
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Not Found",
                  "detail": "Tool not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/tools/{tool_name}/mock": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Execute mock response",
        "description": "Execute mock response for a tool with given parameters.",
        "operationId": "execute_mock_api_v0_tools__tool_name__mock_post",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          },
          {
            "name": "repo_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Repository name",
              "default": "default",
              "title": "Repo Name"
            },
            "description": "Repository name"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MockExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_MockExecutionResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Not Found",
                  "detail": "Tool not found"
                }
              }
            }
          },
          "400": {
            "description": "Mock execution disabled",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/bad-request",
                  "title": "Bad Request",
                  "detail": "Mock execution is disabled for this tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MockExecutionRequest": {
        "properties": {
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "Parameters to pass to the mock execution"
          }
        },
        "type": "object",
        "title": "MockExecutionRequest",
        "description": "Request model for executing mock response with parameters."
      },
...
      "ResponseStatus": {
        "type": "string",
        "enum": [
          "success",
          "error",
          "partial"
        ],
        "title": "ResponseStatus",
        "description": "Standard response status values."
      },
...
      "StandardResponse_List_ToolSummary__": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "description": "Response status indicator",
            "default": "success"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "description": "HTTP status code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable message about the response"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Response metadata"
          }
        },
        "type": "object",
        "title": "StandardResponse[List[ToolSummary]]",
        "example": {
          "data": {
            "id": 1,
            "name": "Example"
          },
          "message": "Operation completed successfully",
          "meta": {
            "request_id": "req_123",
            "timestamp": "2024-01-01T00:00:00Z",
            "version": "1.0.0"
          },
          "status": "success"
        }
      },
...
    }
  },
...
}